hola xd

+++ src/main/resources/templates/BdsTienda.sql
@@ -24,7 +24,7 @@ apellidos varchar(100),
 contrasenia VARCHAR(100),
 email VARCHAR(100),
 fechaNacimiento date,
-numeroTarjeta INT(16),
+numeroTarjeta BIGINT,
 titularTarjeta VARCHAR (100),
 codigoSeguridad INT(3),
 direccionFacturacion VARCHAR (200)
@@ -33,6 +33,7 @@ select contrasenia from Usuarios WHERE contrasenia = MD5('contrafulanito');
 INSERT INTO Usuarios VALUES(2,"Luis","Martinez",MD5('contrafulanito'),"luis@.com","1994-12-12","12345678","luis","123","garcia plata de osma");
 
 SELECT * FROM Productos;
+
 SELECT * FROM Usuarios;
 
 delete from Usuarios;
